reviewer comments
I took the code and put it into Titanium and did not find any errors. Good use of comments in the code. The phone gadget was the map gadget that was found on page 173 of chapter 7 in the book. The one suggestion that I have for the code is to come up with your own theme for the colors in the app. The colors you used were the ones taken straight from the code in the book and I think it would have been better to come up with your own theme and colors.